% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_rows.R
\name{aggregate_columns}
\alias{aggregate_columns}
\title{Aggregate multiple rows into one.}
\usage{
aggregate_columns(x, columns, colname, mode = "sum", delete = FALSE)
}
\arguments{
\item{x}{Any nonempty dataframe.}

\item{columns}{Names of the columns to be aggregated.}

\item{mode}{Choose "sum" to take sum, or "mean" to take the mean.}

\item{delete}{Choose TRUE to drop the original rows from the dataframe, or TRUE to retain them.}

\item{new_colname}{How to name the new columns}
}
\value{
Returns the dataframe with the new column included.
}
\description{
Multiple columns are added together, or the mean is taken to create a new columns. Used can also choose to delete the original columns.
}
\examples{
aggregate_columns(mtcars, c("vs", "cyl", "mpg"), colname = "meaningless_mean", mode = "mean")
}
