% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astype.R
\name{as.type}
\alias{as.type}
\title{Set column type as stored in variable}
\usage{
as.type(vector, type)
}
\arguments{
\item{vector}{Object to be converted.}

\item{type}{Type in string, e.g., "factor".}
}
\value{
Object x, converted to target type.
}
\description{
For instance to allow to return an object of the same type as was provided.
}
\examples{
type <- "factor"
as.type(mtcars$mpg, "character")
}
