% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_NA.R
\name{build_NA}
\alias{build_NA}
\title{Provide some NA statistics}
\usage{
build_NA(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with multiple rows of statistics on the input dataframe \code{x}
}
\description{
Provide some NA statistics
}
\examples{

}
