% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_duplicates.R
\name{consolidate}
\alias{consolidate}
\title{Utility function for consolidating duplicates.}
\usage{
consolidate(df, summary_parsing, ..., by_cols = NA)
}
\arguments{
\item{df}{Dataframe to consolidate.}

\item{summary_parsing}{Functions to apply in the summarize function.}

\item{...}{Columns by which to group for summarize().}

\item{by_cols}{Alternative argument to allow for passing through the columns by which to group.}
}
\description{
Utility function for consolidating duplicates.
}
\examples{
consolidate_duplicates(mtcars, list(mean_hp = rlang::quo(mean(hp))), cyl)
}
