% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_duplicates.R
\name{consolidate_duplicates}
\alias{consolidate_duplicates}
\title{Utility function for consolidating duplicates.}
\usage{
consolidate_duplicates(df, summary_parsing, ..., by_cols = NA)
}
\arguments{
\item{df}{Dataframe to consolidate.}

\item{summary_parsing}{Functions to apply in the summarize function.}

\item{...}{Columns by which to group for summarize.}

\item{by_cols}{Alternative argument to pass through the columns by which to group.}
}
\description{
Utility function for consolidating duplicates.
}
\examples{
consolidate_duplicates(mtcars, list(mean_hp = rlang::quo(mean(hp))), cyl)
}
