% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_entries.R
\name{delete_entries_containing}
\alias{delete_entries_containing}
\title{Delete columns or rows containing the provided string.}
\usage{
delete_entries_containing(x, string, delete = "columns", exclude = c())
}
\arguments{
\item{x}{Any nonempty dataframe.}

\item{string}{The string to be matched.}

\item{delete}{Should matched columns, or matched rows be deleted. Default is "columns".}

\item{exclude}{Exclude specific column named here.}
}
\value{
Returns the dataframe without all columns that contain the string.
}
\description{
The provided dataframe is returned with out any of the columns or rows that contain the specified string.
}
\examples{
delete_entries_containing(mtcars, "a", exclude = "am")
delete_entries_containing(mtcars, "Toyota", delete = "columns")
}
