% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_empty_observations.R
\name{exclude_empty_observations}
\alias{exclude_empty_observations}
\title{Exclude empty observations}
\usage{
exclude_empty_observations(df, columns)
}
\arguments{
\item{df}{Dataframe}

\item{columns}{Columns to be considered}
}
\value{
The dataframe, where observations are removed if they are completely empty or 0
}
\description{
Exclude empty observations
}
\examples{
df <- data.frame(a = c(0, 2, 0),
b = c(0, 3, 0),
c = c(NA, 1, 2))
exclude_empty_observations(df, c("a", "b", "c"))
}
