% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupyter_styling.R
\name{jupyter_styling}
\alias{jupyter_styling}
\title{Adjust print options for a dataframe in jupyter lab using kable.}
\usage{
jupyter_styling(df, ...)
}
\arguments{
\item{df}{The dataframe to print}

\item{...}{Arguments to be passed through to \link[kableExtra]{kable_styling}.}
}
\description{
Workaround for using kable styling in jupyter lab.
}
\examples{
jupyter_styling(mtcars, font_size=14)
}
