% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplots.R
\name{plot_2density}
\alias{plot_2density}
\title{2d density plot}
\usage{
plot_2density(df, col_y, cols = NULL, alpha = 0.2)
}
\arguments{
\item{df}{Any non-empty dataframe.}

\item{col_y}{Column with the variable to be put on the y-axis.}

\item{cols}{A number of columns from the dataframe, for each of which a 2d density plot will be created. By default, will use all columns from the dataframe, except for the y-axis column.}

\item{alpha}{Transparency.}
}
\value{
Multiple 2d density plot, by default one for each column in the provided dataframe, with the selected y column on the y-axis.
}
\description{
Creates a 2d density plot of the selected variables.
}
\examples{
plot_2density(mtcars, "qsec")
}
\references{
https://www.edx.org/course/principles-of-machine-learning-r-edition
}
