% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplots.R
\name{plot_dist}
\alias{plot_dist}
\title{Kernel density plot}
\usage{
plot_dist(df, numcols = colnames(df))
}
\arguments{
\item{df}{Any non-empty dataframe.}

\item{numcols}{Names of the columns to be used. By default, will use all numeric columns.}
}
\value{
Will print one density plot for the columns in \code{df}, or those selected through the parameter \code{numcols}.
}
\description{
Visualization of the kernel density estimation. A density plot is returned for each numeric column in the provided dataframe.
}
\examples{
plot_dist(cars)
}
\references{
https://www.edx.org/course/principles-of-machine-learning-r-edition
}
