% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplots.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot histogram}
\usage{
plot_hist(df, numcols = colnames(df), bins = 10)
}
\arguments{
\item{df}{Any non-empty dataframe.}

\item{numcols}{The names of the columns to be visualized. By default will use all numeric columns.}

\item{bins}{Number of bins. By default will use 10 bins.}
}
\value{
Prints on histogramfor each numeric column in the provided dataframe.
}
\description{
Returns a histogram for every numeric variable in the provided dataframe.
}
\examples{
plot_hist(cars)
}
\references{
https://www.edx.org/course/principles-of-machine-learning-r-edition
}
