% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplots.R
\name{plot_hist_dens}
\alias{plot_hist_dens}
\title{Plot histogram and kernel density estimation}
\usage{
plot_hist_dens(df, numcols = colnames(df), bins = 10)
}
\arguments{
\item{df}{Any non-empty dataframe.}

\item{numcols}{Names of columns to be used. By default uses all columns.}

\item{bins}{How many bins to be used for the histogram. By default, uses 10 bins.}
}
\value{
Will returned an overlayed combination of density plot and histogram for the columns in \code{df}, or those selected through the parameter \code{numcols}.
}
\description{
A combination of \link[DataAnalysisTools]{plot_hist} and \link[DataAnalysisTools]{plot_dist}.
}
\examples{
plot_hist_dens(cars)
}
\references{
https://www.edx.org/course/principles-of-machine-learning-r-edition
}
