% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_look.R
\name{quick_look}
\alias{quick_look}
\title{Show a preview of a dataframe, even a high-dimensional one. In practical terms, the function is a wrapper around glimpse.}
\usage{
quick_look(df, first = 5, last = 3)
}
\arguments{
\item{df}{Any nonempty dataframe.}

\item{first}{The numbers of columns, starting from the left of the dataframe, to be shown.}

\item{last}{THe number of columns, starting from the right of the dataframe, to be shown.}
}
\value{
Returns the dataframe without all columns that contain the string.
}
\description{
The first n columns and last m columns are shown. This avoids clutter in a workbook e.g., when a dataframe has > 100 columns.
}
\examples{
quick_look(mtcars)
}
