% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_company_suffixes.R
\name{remove_company_suffixes}
\alias{remove_company_suffixes}
\title{Remove company suffixes from a string.}
\usage{
remove_company_suffixes(string)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one.}
}
\value{
A character vector.
}
\description{
Remove company suffixes from a string.
}
\examples{
remove_company_suffixes(c("Enterprise Products Operating LLC", "Mercedes-Benz AG"))
}
