% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astype.R
\name{as.type}
\alias{as.type}
\title{Set column type as stored in variable}
\usage{
as.type(x, type)
}
\arguments{
\item{x}{Object to be converted.}

\item{type}{Type in string, e.g., "factor".}
}
\value{
Object x, converted to target type.
}
\description{
Set column type as stored in variable
}
\examples{
type <- "factor"
}
