% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_NA.R
\name{build_NA}
\alias{build_NA}
\title{Provide statistics on missing data in dataframes.}
\usage{
build_NA(x)
}
\arguments{
\item{x}{Any nonempty data.frame.}
}
\value{
A dataframe with origin of data, number of NAs, number of observations, and % missing data of \code{x}. The output is designed so that that mutliple of the returned dataframes can be concatenated to give a quick overview over the missing data in multiple dataframes. Function returns name of the dataframe, and for each variable (column) the number of NAs, total number of observations (rows), and percentage of missing data. Missing data must be formated to R NA values.
}
\description{
Provide statistics on missing data in dataframes.
}
\examples{
NA1 <- build_NA(mtcars)
NA2 <- build_NA(cars)
NA_statistics <- rbind(NA1, NA2)
NA_statistics
}
