% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_NA.R
\name{build_NA}
\alias{build_NA}
\title{Provide statistics on misssing data in dataframes.}
\usage{
build_NA(x)
}
\arguments{
\item{x}{Any onempty data.frame.}
}
\value{
A dataframe with multiple rows of statistics on the input dataframe. The output is designed so that that mutliple of the returned dataframes can be concatenated to give a quick overview over the missing data in multiple dataframes. Function returns name of the dataframe, and for each variable (column) the number of NAs, total number of observations (rows), and percentage of missing data. Missing data must be formated to R NA values.  \code{x}
}
\description{
Provide statistics on misssing data in dataframes.
}
