% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_duplicates.R
\name{consolidate_duplicates}
\alias{consolidate_duplicates}
\title{Consolidate duplicate entries}
\usage{
consolidate_duplicates(df, summary_parsing, ..., by_cols = NA)
}
\arguments{
\item{df}{Dataframe to consolidate.}

\item{summary_parsing}{Functions to apply in the summarize function.}

\item{...}{Columns by which to group for (summarize).}

\item{by_cols}{Alternative argument to allow for passing through the columns by which to group.}
}
\description{
Consolidates the columns for which there are summary functions provided. 
For columns without summary functions, non-duplicated entries retain their 
original value, and duplicated entries return NA.
}
\examples{
consolidate_duplicates(mtcars, list(wt = rlang::quo(mean(wt))), hp)
}
